functions/slackbot-listener/requirements.txt
google-api-python-client
google-cloud-container
google-cloud-pubsub
flask
slackclient